+++ modules/generator/src/test/java/mada/tests/e2e/dto/jspecify/records/dto/Pet.java
@@ -6,31 +6,35 @@
  * Contact: openapi API contact email
  */
 
-package mada.tests.e2e.dto.jspecify.dto;
+package mada.tests.e2e.dto.jspecify.records.dto;
 
 import java.util.Objects;
-import java.util.UUID;
 import javax.json.bind.annotation.JsonbProperty;
 import javax.validation.constraints.NotNull;
 import org.eclipse.microprofile.openapi.annotations.media.Schema;
 import org.jspecify.annotations.Nullable;
 
 /**
- * AdresseIdentifikationDto
+ * Pet
  */
 @javax.annotation.processing.Generated(value = "dk.mada.jaxrs.Generator")
-public record AdresseIdentifikationDto(
-  @JsonbProperty("adgangsadresseId")
+public record Pet(
+  @JsonbProperty("id")
   @Schema(required = true)
   @NotNull
-  UUID adgangsadresseId,
+  Long id,
 
-  @JsonbProperty("adresseId")
-  @Schema(required = true, nullable = true)
+  @JsonbProperty("name")
+  @Schema(required = true)
+  @NotNull
+  String name,
+
+  @JsonbProperty("tag")
   @Nullable
-  UUID adresseId) {
+  String tag) {
 
-    public AdresseIdentifikationDto {
-      Objects.requireNonNull(adgangsadresseId, "Property adgangsadresseId is required, cannot be null");
+    public Pet {
+      Objects.requireNonNull(id, "Property id is required, cannot be null");
+      Objects.requireNonNull(name, "Property name is required, cannot be null");
     }
 }
