+++ modules/generator/src/test/java/mada/tests/e2e/dto/jspecify/records/test.properties
